refactor: moving options helpers 'Is Quorum <...>' to llmq/options
